test(pyspark): clean up duplicated array tests
